Build SkedDataReader to save and load SkedData class
Unit testing for this push is partially broken because it requires an
unwriteable file, which is impossible to pull consistently, and an
unreadable file, which cannot even be pushed.  These tests use the
Assume class to skip the test if `user_UNREADABLE.json` and/or
`user_UNWRITEABLE.json` do not exist.