UI Flow cleanup, resolves #15
The NavigationTargetPane had to be changed to an interface because the
ScrollPane does not interact well with grandchildren panes.  The
FlowPane doesn't work well with the ScrollPane unless it is a direct
child and isn't wrapped in a regular "Pane" class.  It is now the
responsibility of the NavigationTarget to populate the getNode() method
with a ScrollPane if it is needed.