Remove defensive copy from AssignmentListOrganizer constructor
The original decision for making a defensive copy was made under the
assumption that organizing would use an already existing list.  However,
this made reusing the organizer impossible, so a defensive copy was
being made for that as well.  In other words, it makes more sense to
just copy the list right before it is organized than to have two copies
made.

Furthermore, making a defensive copy made reusing the organizer
difficult, as it would not be up-to-date with all of the assignments
when organize() was called.